% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA_p-function.R
\name{APA_p}
\alias{APA_p}
\title{APA style options for p values}
\usage{
APA_p(stat, digit = 3, as_range = FALSE, rmd_format = TRUE)
}
\arguments{
\item{stat}{A number, supplied as a numeric or character value.}

\item{digit}{How many significance digits (defaults to 3)? Passed to \code{\link[formattable]{formattable}} with options \code{format = "f"} and \code{digits = digit}.}

\item{as_range}{Should values be returned as exact numbers (rounded, with no leading zero)? Defaults is \code{TRUE}. \code{FALSE} returns numbers with greater than/ less than indicators and no leading zero.}

\item{rmd_format}{Should p value be return with a Latex/rMarkdown-friendly label and equal sign prepended (e.g., \strong{"\\\\textit{p} $=$ .012"}). Default is \code{TRUE}.}
}
\value{
\code{stat} as a string with APA formatting and/or other options applied.
}
\description{
\code{APA_p} takes a number (numeric or character) and returns it with appropriate APA formatting, with options for print-friendly labelling. For p values less than .001, APA recommends using p < .001. Otherwise, numbers can be returned as exact values (rounded and without a leading zero; e.g., \strong{p = .003}) for reporting p values in text, or as values in a range (for tables, etc.; e.g., \strong{p < .05}).
}
\examples{

# lapply(runif(n = 100, min = 0.00003, max = .99), APA_p, rmd_format = TRUE)
}
