% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA_sem-function.R
\name{APA_sem}
\alias{APA_sem}
\title{Return a Std. Error of the Mean (SEM) value in APA style}
\usage{
APA_sem(stat, rmd_format = TRUE, digit = 3)
}
\arguments{
\item{stat}{A number, supplied as a numeric or character value.}

\item{rmd_format}{Either \code{TRUE} (the default) to return the number with a Latex/rMarkdown-friendly test label and equal sign prepended (e.g., \strong{"\\\\textit{MS$_e$} $=$ 12.34"}); or, \code{FALSE}, to return just the number with rounding.}

\item{digit}{How many significant digits (defaults to 3)? Passed to \code{\link[formattable]{formattable}} with options \code{format = "f"} and \code{digits = digit}.}
}
\value{
A string.
}
\description{
\code{APA_sem} takes a number (numeric or character) and returns it with appropriate APA formatting, with options for print-friendly labelling.
}
\examples{

# lapply(runif(n = 100, min = 0.10, max = 10.0), APA_sem)
}
