% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA_stat-function.R
\name{APA_stat}
\alias{APA_stat}
\title{Format a statistic}
\usage{
APA_stat(stat, pre_stat = NULL, post_stat = NULL, snip = FALSE,
  digit = 2, ...)
}
\arguments{
\item{stat}{A number, supplied as a numeric or character value.}

\item{pre_stat}{A string to be prepended to you number. Default is \code{NULL}.}

\item{post_stat}{A string to be appended to your number. Default is \code{NULL}.}

\item{snip}{Do you want to snip a leading zero from a number bounded by 1? Defaults to \code{FALSE}.}

\item{digit}{How many significance digits (defaults to 2, which is suitable for many APA applications)? Passed to \code{\link[formattable]{formattable}}.}

\item{...}{Optional arguments to be passed to \code{\link[formattable]{formattable}} or \code{paste0}}
}
\value{
\code{stat} as a string with APA formatting and/or other options applied.
}
\description{
\code{APA_stat} allows you to append your own labels to a statistic.
}
\examples{

# lapply(runif(n = 100, min = 0.00003, max = .99), APA_p, rmd_format = TRUE)
}
