% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_stars-function.R
\name{p_stars}
\alias{p_stars}
\title{Return p-value significance star(s) from a number.}
\usage{
p_stars(p, show_ns = TRUE)
}
\arguments{
\item{p}{A number for which a p-value star should be returned:

 \describe{
  \item{\strong{***}}{0 < \emph{p} < 0.001}
  \item{\strong{**}}{0.001 < \emph{p} < 0.01}
  \item{\strong{*}}{0.01 < \emph{p} < 0.5}
  \item{\strong{.}}{0.05 < \emph{p} < 0.1}
}}

\item{show_ns}{Show "n.s." for non-significant values? Default = TRUE}
}
\value{
A string.
}
\description{
Uses the standard range of intervals to determine the nubmer of stars.
}
