% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA_p-function.R
\name{APA_p}
\alias{APA_p}
\title{APA style options for \emph{p} values}
\usage{
APA_p(stat, digit = 3, not_range = TRUE, rmd_format = TRUE)
}
\arguments{
\item{stat}{A number, supplied as a numeric or character value.}

\item{digit}{How many significance digits (defaults to 3)?

Passed to \code{\link[formattable]{formattable}} with options \code{format = "f"} and \code{digits = digit}.}

\item{not_range}{Should full \emph{p} values be returned instead of significance ranges? Default is \code{TRUE}. \code{FALSE} returns numbers with greater than/ less than indicators and no leading zeros.
See \strong{P values}.}

\item{rmd_format}{Should \emph{p} value be return with a Latex/rMarkdown-friendly label and equal sign prepended (e.g., \strong{"\\\\textit{p} $=$ .012"}).
Default is \code{TRUE}.}
}
\value{
\code{stat} as a string with APA formatting and/or other options applied.
}
\description{
\code{APA_p} takes a number (numeric or character) and returns it with appropriate APA formatting, with options for print-friendly labelling.
}
\section{P values}{

For p values less than .001, \strong{\href{http://blog.apastyle.org/apastyle/statistics/}{APA}} recommends using \strong{\emph{p} < .001}. Otherwise, full \emph{p} values should be used, rounded to two or three decimal places, and without a leading zero (e.g., \strong{\emph{p} = .003}) when reporting \emph{p} values in text, or as values in a range (for tables, etc.; e.g., \strong{\emph{p} < .05}). See \code{\link{snip}} for details about leading and trailing zeros.
}

\examples{
# lapply(runif(n = 100, min = 0.00003, max = .99), APA_p, rmd_format = TRUE)
}
