% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snip-function.R
\name{snip}
\alias{snip}
\title{Remove leading zero and/or trailing zero from a number.}
\usage{
snip(x, lead_zero = FALSE, trail_zero)
}
\arguments{
\item{x}{A numeric value, coercible to a character.}

\item{lead_zero}{Remove a zero to the left of the decimal for bounded numbers? Default is \code{FALSE}.}

\item{trail_zero}{= Optional. Set minimum number of digits to the right of the decimal to keep, and then remove the next digit if it is a zero, or else preserve it; if \code{trail_zero} > the actual number of digits  of \code{x}, returns \code{x} unchanged.}
}
\value{
\code{x} as a string with leading and/or trailing zero removed
}
\description{
Useful for APA-style reporting of numbers bounded by 1/-1 (e.g., converts 0.23 to .23); can aslo be used to drop a trailing zero (e.g., converts 2.270 to 2.27) . NOTE: that this converts the number to a character vector and is not guaranteed to produce the appropriate result in all cases.
}
\examples{

# Drop leading zero
cat(snip("0.2202",lead_zero = TRUE))

# Drop leading zero, keep a minimum of 2 digits to the right, or drop if 0
cat(snip("0.2202",lead_zero = TRUE,2))
cat(snip("0.2230",lead_zero = TRUE,2))

}
